basic Schindler page (85%) Done
the basic version of the schindler page is "Almost" done. Most of the components are made and should be functional. The only thing to do is making the Programming Board and the LCD Screen components, and adding the commands we will be using in the modu-tool. But for now, its done.